@@ -3,7 +3,7 @@ version: 1.0.0+1
 description: Uma aplicação de troca de habilidades - Ensina o que sabes, aprende algo novo!
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 
 dependencies:
   flutter:
